Move current_branch() into prompt module for now
Since this appears to be the only user, it seems sensible to keep it close
to the related code for the time being. If broader Git functionality is
needed in other places, it can always be extracted again.
