+++ src/lib.rs
@@ -17,7 +17,6 @@ mod env;
 mod errors;
 mod evaluate;
 mod format;
-mod git;
 mod object;
 mod parser;
 mod plugin;
