+++ src/prompt.rs
@@ -1,4 +1,6 @@
-use crate::git::current_branch;
+use git2::{Repository, RepositoryOpenFlags};
+use std::ffi::OsString;
+
 use crate::prelude::*;
 
 pub struct Prompt;
@@ -17,5 +19,22 @@ impl Prompt {
                 None => "".to_string(),
             }
         )
+    }
+}
+
+pub fn current_branch() -> Option<String> {
+    let v: Vec<OsString> = vec![];
+    match Repository::open_ext(".", RepositoryOpenFlags::empty(), v) {
+        Ok(repo) => {
+            let r = repo.head();
+            match r {
+                Ok(r) => match r.shorthand() {
+                    Some(s) => Some(s.to_string()),
+                    None => None,
+                },
+                _ => None,
+            }
+        }
+        _ => None,
     }
 }
